Fixed a bug due to which Synfig does not start if the username contains non-Latin characters (Windows)
Finally, I found a proper solution for the problem with file paths
containing non-Latin characters in Windows.

The main problem was that we are working with UTF-8, while on Windows
filenames use UTF-16. To work with files, we use the Glibmm wrapper,
which performs correct path conversion, but in some cases we need
standard cpp `std::fstream` objects.

Unfortunately, the standard implementation does not support `std::wstring`
paths. C++17 introduced support through the use of `std::filesystem::path`,
but since we continue to support c++11, this option is not suitable for us.

Luckily, both MSVC and MinGW support a non-standard constructor overload
using wchar_t, and on Windows we can use it. All that is required is to convert
the path to UTF-16 for Windows and leave it as is for other operating systems.

For this, a special method `FileSystemNative::path(...)` has been added, which
does all the necessary work of converting the path when necessary.